Section 4.1 in Book
Did section 4.1. Changed files/got ride of TXT. Added my name and date to their spots in the files. Also updated links so that they linked to the correct spots. Also went through and saved each file so beautify would make each code look pretty. Added CSS styles to the figure, made it align to the center of the body element and set the width of it. Made the image in the figure display block and made it a width of 100% of the figure element. Then set the figure caption to have a white background, have a different font family then the rest of the text, made the text appear italic and align center, and put a padding of 10px on top and bottom. Then set the background image of html tag to be a custom image. Set the background image of the article to display a small image in the bottom left and right corner of the page and then put a different background for the article, but because of it's size it covers the body of the page. Then put a border on the body and the aside. Then set the aside border to have curved corners. Then set the figure to have a border with a custom image as the border to make it look like a torn out page.