Section 4.2 in Book
Changed the opacity of the figure element. Gave the Article a box shadow to the left and to the right. Gave the Heading 1 element a text shadow. Then gave the body a shadow that goes inward and not outward. Then gave the aside a shadow as well. Made the aside change from a white to a green starting in a circle in the middle of the element to the edge of the element. Then gave the footer a linear gradient starting from the top left (Black) of the element to the bottom right (White) of the element, but going over the green spectrum.