Section 4.3 in Book
Linked the style sheets to tb_genta. Set the perspective of the article to 600px to set up for the figure to have a 3d  transformation. then transformed each figure on the page to have a 3d transformation and also to have a box shadow. Then gave each figure a filter, figure1 got a sepia, figure2 got grayscale, figure3 got saturation and contrast. On the index page put image map so that each face on the image is linked to their respective page.